+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 -   Added support for parameterized Gherkin steps, per #431
 
+### Changed
+
+-   **BREAKING** changed `withTreeNode` to `withNode` in Message class in Handlers.ts
+
 ## [0.16.0] - 2017-03-19
 
 [0.16.0]: https://github.com/atomist/rug/compare/0.15.1...0.16.0